Allow reading token from the config file, env variable & command output.
Some users might prefer to keep their passwords/api_tokens in a keychain
so as not to expose them unencrypted in an ENV variable or a config
file.

This commit introduces an `api_token_cmd` config key (and a
corresponding `JIRA_API_TOKEN_CMD` environment variable) which,
if set, specifies a command to run whose output will provide the value
of the api token.

This can be set to, e.g.,

`kwallet-query -r my.jira.server.com kdewallet`

for the KDE KWallet, or an appropriate command for other keychains or
password managers.

See also Issue #247.