Respect `WT_SESSION` variable.
This is in accordance to [this comment](https://github.com/microsoft/terminal/issues/1040#issuecomment-496691842). As discussed further in the issue, in an ideal world, we wouldn't check "user agents" of terminals one-by-one, but instead would do proper feature probing like described [here](https://github.com/termstandard/colors).

But anyway: This is a good first step to enhance Windows support.