fix: regression introduced by new OriginalEstimate attribute
The recent changes in PR #748 caused issues with jira-cli when used with Jira Enterprise On-Prem (now referred to as DataCenter). All requests started failing with the following error message:

```
Error:
  - timetracking: Field 'timetracking' cannot be set. It is not on the appropriate screen, or unknown.
```

The issue arises because, even if the `--original-estimate` option is not specified, the timetracking field is included in the JSON query.
This PR addresses the issue by ensuring that the timetracking field is only included in the JSON query if the `--original-estimate` option is set.
I have tested the fix, and all tests are passing.

Please review the changes and let me know if there are any concerns.