fix: timetracking field always marshalled
Currently, the field for timetracking is always marshalled to JSON even if the "OriginalEstimate" field is an empty string.

This is because of the way the field is initialized as a struct that will never be nil and therefore never caught by the 'omitempty' tag.

This changes the TimeTracking field to be a pointer, which is nil by default, so that if OriginalEstimate isn't set, the timetracking field won't be in the JSON sent to Jira.

This is a problem if the issue type being created is one that doesn't support a timetracking field.