feat: add cf-access as supported auth type
Adds a new auth type `cf-access` which allows using this tool against
servers behind cloudflare access. Defines the new auth type, create a
new RoundTripper to handle the login / token fetch flow, and set it as
the client transport if enabled. Fetch the needed token on jira client
creation if the auth type is set to `cf-access`.