fix: case insensitive sub-task comparison
Within my cloud installation, the jira init command pulls the handle and name just fine but my administrator configured the name as "Sub-task" and the handle as "Sub-Task" which fails the equal check and causes failures with the parent key assignment. This PR addresses this by using an EqualFold so case insensitive matching is done for the issue type check.