fix: use `tview.Escape` with standard square brackets (#828)
* fix: use `tview.Escape` with standard square brackets

The original implementation of `prepareTitle`, as well as the date ranges for sprints, used nonstandard brackets `⦗⦘` as replacements for square brackets `[]` in a hack to work around `tview` interpreting square brackets as format tags. This was causing visual problems for some powerline fonts.

The [documentation](https://pkg.go.dev/github.com/rivo/tview) for `tview` mentions that bracketed strings may be escaped by typing `[text[]` instead of `[text]`, and the package provides `tview.Escape(str)` to perform that conversion.

This change switches to using the escape function. I tried to minimize the amount of things I was touching, which also means that the expected strings in tests are a bit ugly.

* remove extraneous formatting changes