Add a --board parameter to sprint list
Add an optional `--board` parameter that takes in input a board id that overrides the
parameter specified in the config file. This avoids the need to edit the config file
in case a change of board is needed.

Also implement dynamic retrieval of the board name in case the board parameter is used.

Parameter is added only to `sprint list` command as it is the only one requiring a
board at the moment, but it can be extended in the future.